Move COA related models to Models\ChartOfAccounts
